+++ tests/testthat/test_predict.R
@@ -383,3 +383,19 @@ test_that("predictInterval makes predictions without observed outcome", {
   expect_is(testPreds2, "data.frame")
   expect_is(testPreds3, "data.frame")
 })
+
+context("Input validation checks")
+
+test_that("dplyr objects are successfully coerced", {
+  set.seed(101)
+  library(dplyr)
+  m1 <- lmer(Reaction ~ Days + (1 | Subject), sleepstudy)
+  predData <- sleepstudy %>% group_by(Subject) %>% summarize(Days = mean(Days))
+  expect_warning(predictInterval(m1, newdata = predData), "newdata is tbl_df or tbl object from dplyr package and has been
+              coerced to a data.frame")
+  preds2 <- predictInterval(m1, newdata = predData, n.sims=2000)
+  expect_is(preds2, "data.frame")
+  predData2 <- as.data.frame(predData)
+  preds1 <- predictInterval(m1, newdata = predData2, n.sims=2000)
+  expect_true(sum(preds1$fit - preds2$fit) > -50 & sum(preds1$fit - preds2$fit) < 50)
+})